+++ CHANGELOG.md
@@ -16,6 +16,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Tests for "Number derivation".
 - Tests for "Simple auto-scaling".
 - Tests for "Horse-racing hyperduals".
+- Tests for "Balanced ternary computer: encode".
 
 ## [1.8.0] - 2022-04-30
 ### Added